Ruby: fixed contents of SCRIPT_NAME.
Having the basename of the script pathname was incorrect.  While
we don't have something more accurate, the best thing to do is to
have it empty (which should be most of the time).

Fixes (git) commit 0032543fa65f454c471c968998190b027c1ff270,
which was 'Ruby: added the Rack environment parameter "SCRIPT_NAME".'.